backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.33.0",
    "python-dotenv>=0.20.0",
    "pytest>=7.4.0",
]
